Refactor code to new Parser Classes
Moved prepare methods from the CommandManager class to their respective
Parser classes.
