+++ src/main/java/seedu/duke/parser/AddExerciseParser.java
@@ -1,2 +1,33 @@
-package seedu.duke.parser;public class AddExerciseParser {
+package seedu.duke.parser;
+
+import seedu.duke.command.Command;
+import seedu.duke.command.IncorrectCommand;
+import seedu.duke.command.exercise.AddExerciseCommand;
+import seedu.duke.exception.GetJackDException;
+
+public class AddExerciseParser extends Parser {
+
+    public AddExerciseParser(String userInputString) {
+        this.userInputString = userInputString;
+    }
+
+    private Command prepareAddExercise(String commandArgs) {
+        try {
+            String[] exerciseArgs = getExerciseArgs(commandArgs);
+            String exerciseName = exerciseArgs[1].trim();
+            int workoutIndex = parseArgsAsIndex(exerciseArgs[0]);
+            int sets = parseArgsAsIndex(exerciseArgs[2]);
+            int reps = parseArgsAsIndex(exerciseArgs[3]);
+
+            return new AddExerciseCommand(workoutIndex, exerciseName, sets, reps);
+        } catch (GetJackDException e) {
+            return new IncorrectCommand(MESSAGE_INVALID_COMMAND + AddExerciseCommand.MESSAGE_USAGE);
+        }
+    }
+
+    @Override
+    public Command parseInput() {
+        String commandArgs = getCommandArguments(userInputString);
+        return prepareAddExercise(commandArgs);
+    }
 }
